requirements.txt
websocket-client
requests
websocket-server